tvplanit: Add missing BaseDatastore and FlexDatastore to package. Some cosmetics.
git-svn-id: https://svn.code.sf.net/p/lazarus-ccr/svn@4682 8e941d3f-bd1b-0410-a28a-d453659cc2b4

+++ components/tvplanit/source/vpbaseds.pas
@@ -35,7 +35,7 @@ interface
 
 uses
   {$IFDEF LCL}
-  LMessages,LCLProc,LCLIntf,
+  LMessages, LCLProc, LCLIntf,
   {$ELSE}
   Windows,
   {$ENDIF}
@@ -228,7 +228,7 @@ type
 
     procedure DeregisterAllWatchers;
     procedure DeregisterWatcher (Watcher : THandle);
-    function GetNextID(TableName: string): int64; virtual; abstract;
+    function GetNextID(TableName: string): Integer; virtual; abstract;
     property Resources: TVpResources read FResources;
     procedure Load; virtual;
     procedure LoadEvents; virtual; abstract;
